removed dtype from LayerNorm

+++ nemo/utils/export_utils.py
@@ -280,7 +280,7 @@ try:
         if isinstance(n, FusedLayerNorm) or isinstance(n, MixedFusedLayerNorm):
             mod = nn.LayerNorm(n.normalized_shape, eps=n.eps, elementwise_affine=n.elementwise_affine,).to(dev)
         elif isinstance(n, FastLayerNorm):
-            mod = nn.LayerNorm(n.weight.shape, eps=n.epsilon, elementwise_affine=True, dtype=torch.float,).to(dev)
+            mod = nn.LayerNorm(n.weight.shape, eps=n.epsilon, elementwise_affine=True,).to(dev)
 
         n_state = n.state_dict()
         mod.load_state_dict(n_state)
